requirements.txt
numpy<2.0

# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3
aiofiles==23.2.1



# AI & ML
groq==0.4.1
anthropic==0.18.1
tiktoken==0.5.2

# OCR & Image Processing
pytesseract==0.3.10
easyocr==1.7.1
Pillow>=10.4.0
opencv-python==4.9.0.80

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.4
pdf2image==1.17.0

# Audio Processing
faster-whisper==1.0.3
pydub==0.25.1
soundfile==0.12.1

# YouTube
youtube-transcript-api==0.6.2

# NLP & Text Processing
nltk==3.8.1
textstat==0.7.3

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.1
httpx==0.26.0
redis==5.0.1
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Logging & Monitoring
loguru==0.7.2
python-json-logger==2.0.7